// functions/api/_lib/analytics/types.ts
export type AnalyticsDateRange = {
  startDate: string;
  endDate: string;
};